+++ libs/cardano-ledger-core/src/Cardano/Ledger/State.hs
@@ -3,6 +3,7 @@ module Cardano.Ledger.State (
   module Cardano.Ledger.State.ChainAccount,
   module Cardano.Ledger.State.CertState,
   module Cardano.Ledger.State.UTxO,
+  module Cardano.Ledger.State.LeiosCommittee,
   module Cardano.Ledger.State.PoolDistr,
   module Cardano.Ledger.State.SnapShots,
   module Cardano.Ledger.State.Stake,
@@ -14,6 +15,7 @@ import Cardano.Ledger.State.Account
 import Cardano.Ledger.State.CertState
 import Cardano.Ledger.State.ChainAccount
 import Cardano.Ledger.State.Governance
+import Cardano.Ledger.State.LeiosCommittee
 import Cardano.Ledger.State.PoolDistr
 import Cardano.Ledger.State.SnapShots
 import Cardano.Ledger.State.Stake